JTS context: makeRectFromRectangularPoly; moved from JtsWKTReader. * GeoJSON polygon reading will convert to rect if possible, and in so honor the datelinerule setting * folded GeoJSON specific tests into one place -- GeneralGeoJSONTest. * reverted some test data number variations
