+++ src/test/java/com/spatial4j/core/io/GeneralGeoJSONTest.java
@@ -17,6 +17,7 @@
 
 package com.spatial4j.core.io;
 
+import com.spatial4j.core.shape.Shape;
 import org.junit.Assert;
 import org.junit.Before;
 import org.junit.Test;
@@ -62,5 +63,87 @@ public class GeneralGeoJSONTest extends GeneralReadWriteShapeTest {
   @Test
   public void testCircle() {
     assertRoundTrip(wkt("BUFFER(POINT(-10 30), 40)"), false);
+  }
+
+  //
+  // Below ported from GeoJSONReadWriteTest:
+  //
+
+  @Test
+  public void testParsePoint() throws Exception {
+    Shape v = reader.read(pointText());
+    assertTrue(point().equals(v));
+  }
+
+  @Test
+  public void testEncodePoint() throws Exception {
+    assertEquals(pointText(), writer.toString(point()));
+  }
+
+  @Test
+  public void testParseLineString() throws Exception {
+    assertTrue(line().equals(reader.read(lineText())));
+  }
+
+  @Test
+  public void testEncodeLineString() throws Exception {
+    assertEquals(lineText(), writer.toString(line()));
+  }
+
+  @Test
+  public void testParsePolygon() throws Exception {
+    assertTrue(polygon1().equals(reader.read(polygonText1())));
+    assertTrue(polygon2().equals(reader.read(polygonText2())));
+  }
+
+  @Test
+  public void testEncodePolygon() throws Exception {
+    assertEquals(polygonText1(), writer.toString(polygon1()));
+    assertEquals(polygonText2(), writer.toString(polygon2()));
+  }
+
+  @Test
+  public void testParseMultiPoint() throws Exception {
+    assertTrue(multiPoint().equals(reader.read(multiPointText())));
+  }
+
+  @Test
+  public void testEncodeMultiPoint() throws Exception {
+    assertEquals(multiPointText(), writer.toString(multiPoint()));
+  }
+
+  @Test
+  public void testParseMultiLineString() throws Exception {
+    assertTrue(multiLine().equals(reader.read(multiLineText())));
+  }
+
+  @Test
+  public void testEncodeMultiLineString() throws Exception {
+    assertEquals(multiLineText(), writer.toString(multiLine()));
+  }
+
+  @Test
+  public void testParseMultiPolygon() throws Exception {
+    assertTrue(multiPolygon().equals(reader.read(multiPolygonText())));
+  }
+
+  @Test
+  public void testEncodeMultiPolygon() throws Exception {
+    assertEquals(multiPolygonText(), writer.toString(multiPolygon()));
+  }
+
+  @Test
+  public void testEncodeRectangle() throws Exception {
+    assertEquals(rectangleText(), strip(writer.toString(rectangle())));
+  }
+
+  @Test
+  public void testParseGeometryCollection() throws Exception {
+    assertEquals(collection(), reader.read(collectionText()));
+  }
+
+  @Test
+  public void testEncodeGeometryCollection() throws Exception {
+    assertEquals(collectionText(), strip(writer.toString(collection())));
   }
 }
